requirements.txt
Django~=3.2.5
celery~=5.1.2
decouple~=0.0.7
jdatetime~=3.6.2
djangorestframework~=3.12.4